Use long time zone names in NextRuns display
Changed the time zone display format from 'short' to 'long' in the NextRuns component to provide more descriptive time zone information.

+++ src/components/NextRuns.tsx
@@ -50,7 +50,7 @@ export const NextRuns: FC<NextRunsProps> = ({ cron, disabled }) => {
 						hour: "2-digit",
 						minute: "2-digit",
 						second: "2-digit",
-						timeZoneName: "short",
+						timeZoneName: "long",
 					}),
 				);
 			}
